\name{TrioSGL-package}
\alias{TrioSGL-package}
\docType{package}
\title{Trio Model with a Combination of Lasso and Group Lasso Regularization}
\description{Fit a trio model via penalized maximum likelihood. The model is fit for a path of values of the penalty parameter. This package is based on Noah Simon, et al. (2011) <doi:10.1080/10618600.2012.681250>.}
\details{
\tabular{ll}{
Package: \tab TrioSGL\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2017-12-18\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
Only 1 function:
\code{TrioSGL}
}
\author{Timo Stöcker, Noah Simon, Jerome Friedman, Trevor Hastie, and Rob Tibshirani

Maintainer: Timo Stöcker <timo.stoecker@hhu.de>
}
\references{
 Simon, N., Friedman, J., Hastie, T., Tibshirani, R. (2011). \emph{A Sparse-Group Lasso}. Journal of Computational and Graphical Statistics, 22, 231-245.
}
